codacy fix
